tools/requirements.txt
# Install the SDK for whichever provider you use.
# All three are listed here so CI can install everything upfront;
# unused SDKs add ~5s and cause no harm.
anthropic>=0.40.0
openai>=1.30.0       # used by both 'github' and 'openai' providers
jinja2>=3.1.0
pyyaml>=6.0
